payment-service/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
PyMySQL==1.1.1

